+++ app/scripts/contentscript.js
@@ -47,10 +47,10 @@ function setupInjection () {
 function setupStreams () {
   // setup communication to page and plugin
   const pageStream = new LocalMessageDuplexStream({
-    name: 'contentscript',
-    target: 'inpage',
+    name: 'contentscript2',
+    target: 'inpage2',
   })
-  const pluginPort = extension.runtime.connect({ name: 'contentscript' })
+  const pluginPort = extension.runtime.connect({ name: 'contentscript2' })
   const pluginStream = new PortStream(pluginPort)
 
   // forward communication plugin->inpage
@@ -88,12 +88,12 @@ function setupStreams () {
   )
 
   // connect phishing warning stream
-  const phishingStream = mux.createStream('phishing')
+  const phishingStream = mux.createStream('phishing2')
   phishingStream.once('data', redirectToPhishingWarning)
 
   // ignore unused channels (handled by background, inpage)
-  mux.ignoreStream('provider')
-  mux.ignoreStream('publicConfig')
+   mux.ignoreStream('provider2')
+   mux.ignoreStream('publicConfig2')
 }