+++ app/scripts/platforms/extension.js
@@ -1,6 +1,10 @@
 const extension = require('extensionizer')
 const explorerLink = function (hash, network) {
-  return `http://wanscan.io/tx/${hash}`
+  if (network === '3') {
+    return `http://18.217.171.41/tx/${hash}`
+  } else {
+    return `https://wanscan.io/tx/${hash}`
+  }
 }
 
 class ExtensionPlatform {