+++ ui/app/components/send/tests/send-selectors-test-data.js
@@ -147,7 +147,7 @@ module.exports = {
       },
     ],
     'selectedAddress': '0xd85a4b6a394794842887b8284293d69163007bbb',
-    'provider': {
+    'providerWan': {
       'type': 'testnet',
     },
     'shapeShiftTxList': [