+++ CHANGELOG.md
@@ -180,7 +180,7 @@
 
 ## 4.1.3 2018-2-28
 
-- Ensure MetaMask's inpage provider is named MetamaskInpageProvider to keep some sites from breaking.
+- Ensure MetaMask's inpage provider is named WanmaskInpageProvider to keep some sites from breaking.
 - Add retry transaction button back into classic ui.
 - Add network dropdown styles to support long custom RPC urls